+++ src/router/index.tsx
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react';
 import { BrowserRouter as Router, Switch, Route, useLocation, Redirect } from 'react-router-dom';
 // import { DrawerLayout } from '@pxblue/react-components';
 import { NavigationDrawer } from '../components/navigation/drawer';
-import { Toolbar } from '../components/navigation/Toolbar';
-import { Footer } from '../components/navigation/Footer';
+import { Toolbar } from '../components/toolbars/Toolbar';
+import { Footer } from '../components/toolbars/Footer';
 import { ROUTES } from '../constants/routes';
 import { LessonsPage, LandingPage, PlaceholderPage, ProsPage, TipsPage, BlogsPage, LoginPage } from '../pages';
 import { usePrevious } from '../hooks';
